+++ application/views/ingredients/tips.php
@@ -1,64 +1,27 @@
+<script src="<?php echo site_url(''); ?>includes/js/tips.js"></script>
 <div id="ingredients_tips" class="row">
 	<div class="page-title medium-12 columns">
 		<p>Alchemy Tips</p>
 	</div>
 	<div class="medium-4 large-3 columns">
-		<table class="list">
+		<table id="select_ingredient_tips" class="list">
 			<tr>
-				<th>Single Effects By Price</th>
+				<th>Select Ingredient for Tips</th>
 			</tr>
-			<?php foreach ($effects as $row) { ?>
+			<?php foreach ($ingredients as $row) { ?>
 				<tr>
 					<td>
-						<a href="<?php echo site_url("effects/view/".$row['id']); ?>">
-							<?php echo $row['name']; ?>
-							(<?php echo $row['price']; ?>)
-						</a>
+						<input type="hidden" class="id" value="<?php echo $row['id']; ?>" />
+						<?php echo $row['name']; ?>
 					</td>
 				</tr>
 			<?php } ?>
 		</table>
 	</div>
 	<div class="medium-8 large-9 columns">
-		<table class="list">
-			<tr>
-				<th colspan="3">Best Value Ingredients Combination with Giant's Toe</th>
-				<th>Sell Price</th>
-			</tr>
-			<?php foreach ($with_giant as $row) { ?>
-				<tr>
-					<?php foreach ($row as $key => $col) { ?>
-						<td>
-							<a href="<?php echo site_url("ingredients/view/".$col['id']); ?>">
-								<?php echo $col['name']; ?>
-							</a>
-						</td>
-					<?php if($key > 1) break; } ?>
-					<td>
-						<?php echo array_sum(array_column($row['result'], 'price')); ?>
-					</td>
-				</tr>
-			<?php } ?>
-		</table>
-		<table class="list">
-			<tr>
-				<th colspan="3">Best Value Ingredients Combination without Giant's Toe</th>
-				<th>Sell Price</th>
-			</tr>
-			<?php foreach ($without_giant as $row) { ?>
-				<tr>
-					<?php foreach ($row as $key => $col) { ?>
-						<td>
-							<a href="<?php echo site_url("ingredients/view/".$col['id']); ?>">
-								<?php echo $col['name']; ?>
-							</a>
-						</td>
-					<?php if($key > 1) break; } ?>
-					<td>
-						<?php echo array_sum(array_column($row['result'], 'price')); ?>
-					</td>
-				</tr>
-			<?php } ?>
-		</table>
+		<div id="by_ingredient_cont">
+			<?php echo $by_ingredient; ?>
+		</div>
+		<?php echo $all_ingredients; ?>
 	</div>
 </div>